requirements.txt
anthropic>=0.40.0
feedparser>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
apscheduler>=3.10.0
cryptography>=42.0.0
jinja2>=3.1.0
python-multipart>=0.0.9
pytz>=2024.1
itsdangerous>=2.1.0
slowapi>=0.1.9
